backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.0
python-dotenv>=1.0.0
pymongo>=4.0
gunicorn>=20.1.0
whitenoise>=6.5.0
PyJWT>=2.8.0
certifi>=2023.7.22
dnspython>=2.4.2
